requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bcrypt==5.0.0
brotli==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
distlib==0.4.0
Django==5.2
django-mysql==4.19.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fonttools==4.60.1
freetype-py==2.5.1
fuzzywuzzy==0.18.0
html2image==2.0.7
html5lib==1.1
idna==3.11
kiwisolver==1.4.9
Levenshtein==0.27.1
lxml==6.0.2
matplotlib==3.10.7
multipledispatch==1.0.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.2.4
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.2.1
pillow-avif-plugin==1.5.2
platformdirs==4.4.0
pycairo==1.29.0
pycparser==2.23
pydyf==0.11.0
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.4.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
razorpay==2.0.0
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.34.0
weasyprint==66.0
webencodings==0.5.1
websocket-client==1.9.0
xhtml2pdf==0.2.17
zopfli==0.4.0
